organized classes, added navigation to group
