+++ study_map/lib/data_architecture/accounts.dart
@@ -1,0 +1,14 @@
+class Account {
+  String name; //account name
+  String password; //account password
+
+  Account({required this.name, required this.password}); //constructor
+
+  String getName() {
+    return name;
+  }
+
+  void setPassword(String newpassword) {
+    password = newpassword;
+  }
+}
