+++ study_map/lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'screens/calendar.dart';
 import 'package:study_map/screens/study_map.dart';
+import 'screens/group.dart';
 
 void main() {
   runApp(MaterialApp(
@@ -12,7 +13,8 @@ void main() {
     routes: {
       // '/' : (conetxt) Loading();
       '/study_map': (context) => const StudyMap(),
-      '/calendar': (context) => const CalendarScreen()
+      '/calendar': (context) => const CalendarScreen(),
+      '/group': (context) => const GroupScreen()
     },
   ));
 }
