+++ study_map/lib/screens/group.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:study_map/data_architecture/events.dart';
 
 class GroupScreen extends StatefulWidget {
   const GroupScreen({Key? key}) : super(key: key);
@@ -21,10 +22,10 @@ class _GroupScreenState extends State<GroupScreen> {
 
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Welcome to Flutter',
+      title: 'Study Groups',
       home: Scaffold(
         appBar: AppBar(
-          title: const Text('Welcome to Flutter'),
+          title: const Text('Group Creation'),
         ),
         body: const Center(
           child: Text('Hello World'),
